Add debug prints and update upgrade cost types
Added debug print statements to HealthHandler and SpawnArea for difficulty and timer updates. Changed upgrade cost variables in Upgrades.cs from int to long for better handling of larger values.

+++ Shop/Upgrades.cs
@@ -10,9 +10,9 @@ public partial class Upgrades : VBoxContainer
     [Export] public Button RadiusButton;
     [Export] public Label MoneyLabel;
 
-    private int _damageCost = 5;
-    private int _cooldownCost = 5;
-    private int _radiusCost = 5;
+    private long _damageCost = 5;
+    private long _cooldownCost = 5;
+    private long _radiusCost = 5;
 
     private static int Function(int timesBought)
     {
